requirements.txt
streamlit 
phidata 
google-generativeai 
python-dotenv
PyPDF2
duckduckgo-search
phi-agent